Only delete empty changelog files whose last modified time has not changed (#40)

+++ src/main/java/org/jenkinsci/plugins/workflow/steps/scm/SCMStep.java
@@ -110,6 +110,7 @@ public abstract class SCMStep extends Step {
                     changelogFile = Files.createTempFile(run.getRootDir().toPath(), "changelog", ".xml").toFile();
                 }
             }
+            Long changelogOriginalModifiedDate = (changelogFile != null) ? changelogFile.lastModified() : null;
             SCM scm = createSCM();
             SCMRevisionState baseline = null;
             Run<?,?> prev = run.getPreviousBuild();
@@ -122,10 +123,11 @@ public abstract class SCMStep extends Step {
                 }
             }
             scm.checkout(run, launcher, workspace, listener, changelogFile, baseline);
-            if (changelogFile != null && changelogFile.length() == 0) {
-                // JENKINS-57918/JENKINS-59560: Some SCM plugins don't write anything to the changelog file in some
+            if (changelogFile != null && changelogFile.length() == 0
+                    && changelogOriginalModifiedDate != null && changelogFile.lastModified() == changelogOriginalModifiedDate) {
+                // JENKINS-57918/JENKINS-59560/FakeChangeLogSCM: Some SCMs don't write anything to the changelog file in some
                 // cases. `WorkflowRun.onCheckout` asks the SCM to parse the changelog file if it exists, and
-                // attempting to parse an empty file will cause an error, so we delete empty files before they even get
+                // attempting to parse an empty file will cause an error, so we delete changelog files that were not modified during the checkout before they even get
                 // to `WorkflowRun.onCheckout`.
                 Files.deleteIfExists(changelogFile.toPath());
                 changelogFile = null;
